pyproject.toml
[project]
name = "pdf2pics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mcp[cli]>=1.7.1",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "requests>=2.32.3",
]
